# Conflicts:
#   lib/hooks/useRoles.ts

+++ lib/hooks/useRoles.ts
@@ -1,10 +1,13 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { RoleResponse, RoleType } from "../types/Role";
 import { useAuth } from "../context/AuthContext";
+import { notify } from "../notifications";
+import { useRouter } from "next/navigation";
 
-const useRoles = (page: number = 1, limit: number = 10, search?: string) => {
+const useRoles = (page: number, limit: number, search?: string,) => {
   const { token } = useAuth();
   const queryClient = useQueryClient();
+  const router = useRouter();
 
   const {
     data: rolesData,
@@ -14,12 +17,23 @@ const useRoles = (page: number = 1, limit: number = 10, search?: string) => {
   } = useQuery({
     queryKey: ["roles", page, limit, search],
     queryFn: async () => {
+      if (!token) {
+        notify.error("Error fetching roles", {
+          description: "Please login to continue"
+        })
+        router.push("/login");
+        return;
+      }
       const params = new URLSearchParams({
         page: page.toString(),
         limit: limit.toString(),
       });
       if (search) params.append("search", search);
-      const response = await fetch(`/api/proxy/role-permissions?${params.toString()}`);
+      const response = await fetch(`/api/proxy/role-permissions?${params.toString()}`, {
+        headers: {
+          "Authorization": `Bearer ${token}`,
+        },
+      });
 
       if (!response.ok) {
         throw new Error("Failed to fetch roles");
@@ -39,10 +53,18 @@ const useRoles = (page: number = 1, limit: number = 10, search?: string) => {
       roleName: string;
       permissions: string[];
     }) => {
+      if (!token) {
+        notify.error("Error adding role", {
+          description: "Please login to continue"
+        })
+        router.push("/login");
+        return;
+      }
       const response = await fetch("/api/proxy/role-permissions", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
+          "Authorization": `Bearer ${token}`,
         },
         body: JSON.stringify({
           role_name: roleName,
@@ -72,10 +94,18 @@ const useRoles = (page: number = 1, limit: number = 10, search?: string) => {
       permissions: string[];
       roleId: string;
     }) => {
+      if (!token) {
+        notify.error("Error editing role", {
+          description: "Please login to continue"
+        })
+        router.push("/login");
+        return;
+      }
       const response = await fetch(`/api/proxy/role-permissions/${roleId}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
+          "Authorization": `Bearer ${token}`,
         },
         body: JSON.stringify({
           role_name: roleName,
@@ -101,8 +131,18 @@ const useRoles = (page: number = 1, limit: number = 10, search?: string) => {
     }: {
       roleId: string;
     }) => {
+      if (!token) {
+        notify.error("Error deleting role", {
+          description: "Please login to continue"
+        })
+        router.push("/login");
+        return;
+      }
       const response = await fetch(`/api/proxy/role-permissions/${roleId}`, {
         method: "DELETE",
+        headers: {
+          "Authorization": `Bearer ${token}`,
+        },
       });
 
       if (!response.ok) {
